+++ drivers/misc/mediatek/pmic/mt6359p/v1/pmic_throttling_dlpt.c
@@ -130,7 +130,7 @@ void register_low_battery_notify(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	lbcb_tb[prio_val].lbcb = low_battery_callback;
+	lbcb_tb[(unsigned int)prio_val].lbcb = low_battery_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
 }
@@ -141,7 +141,7 @@ void register_low_battery_notify_ext(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	lbcb_tb_ext[prio_val].lbcb = low_battery_callback;
+	lbcb_tb_ext[(unsigned int)prio_val].lbcb = low_battery_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
 }
@@ -370,7 +370,7 @@ void register_battery_oc_notify(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	occb_tb[prio_val].occb = battery_oc_callback;
+	occb_tb[(unsigned int)prio_val].occb = battery_oc_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
 }
@@ -538,7 +538,7 @@ void register_battery_percent_notify(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	bpcb_tb[prio_val].bpcb = battery_percent_callback;
+	bpcb_tb[(unsigned int)prio_val].bpcb = battery_percent_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
 
@@ -854,7 +854,7 @@ void register_dlpt_notify(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	dlpt_cb_tb[prio_val].dlpt_cb = dlpt_callback;
+	dlpt_cb_tb[(unsigned int)prio_val].dlpt_cb = dlpt_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
